Vytvoreni knihoven pro pohyb a enkodery
