+++ src/main.cpp
@@ -2,6 +2,9 @@
 #include "MeAuriga.h"
 #include "MeRGBLineFollower.h"
 
+#include "movement.h"
+#include "encoder.h"
+
 // mame robota cislo 11
 
 // Levý motor
@@ -97,96 +100,6 @@ volatile int pulseCountL = 0;
 volatile bool stateAL = false;
 volatile bool stateBL = false;
 volatile bool oldStateAL = false;
-// osetreni preruseni od kanalu A enkoderu na pravem motoru - POZOR - enkoder se toci opacne
-void pravyEncoderAInt() {
-  cli(); // zakaz zpracovani dalsich preruseni
-  stateAR = digitalRead(pravyEnkoderA); // nacteni stanu kanalu A
-  if (digitalRead(pravyEnkoderB)){ // vyhodnot pro B=HIGH
-  if ( (oldStateAR - stateAR) < 0) // jaky je typ hrany na kanalu A
-  pulseCountR--; // vzestupna -> pricti puls
-  else
-  pulseCountR++; // sestupna -> odecti puls
-  } else { // vyhodnot pro B=LOW
-  if ( (oldStateAR - stateAR) > 0) // jaky je typ hrany na kanalu A
-  pulseCountR--; // sestupna -> pricti puls
-  else
-  pulseCountR++; // vzestupna -> odecti puls
-  }
-  oldStateAR = stateAR; // uchovej stav kanalu A pro pristi test hrany
-  encMSG = true;
-  sei(); // opet povol zpracovani dalsich preruseni
-}
-
-
-// osetreni preruseni od kanalu A enkoderu na levem motoru
-void levyEncoderAInt() { 
-  cli(); // zakaz zpracovani dalsich preruseni
-  stateAL = digitalRead(levyEnkoderA); // nacteni stanu kanalu A
-  if (digitalRead(levyEnkoderB)){ // vyhodnot pro B=HIGH
-  if ( (oldStateAL - stateAL) < 0) // jaky je typ hrany na kanalu A
-  pulseCountL++; // vzestupna -> pricti puls
-  else
-  pulseCountL--; // sestupna -> odecti puls
-  } else { // vyhodnot pro B=LOW
-  if ( (oldStateAL - stateAL) > 0) // jaky je typ hrany na kanalu A
-  pulseCountL++; // sestupna -> pricti puls
-  else
-  pulseCountL--; // vzestupna -> odecti puls
-  }
-  oldStateAL = stateAL; // uchovej stav kanalu A pro pristi test hrany
-  encMSG = true;
-  sei(); // opet povol zpracovani dalsich preruseni
-}
-
-void levyMotorVpred(int rychlost) {
-  digitalWrite(inMotorLevy1, HIGH);
-  digitalWrite(inMotorLevy2, LOW);
-  analogWrite(pwmMotorLevy, rychlost);
-}
-
-void levyMotorVzad(int rychlost) {
-  digitalWrite(inMotorLevy1, LOW);
-  digitalWrite(inMotorLevy2, HIGH);
-  analogWrite(pwmMotorLevy, rychlost);
-}
-
-void levyMotorStop() {
-  analogWrite(pwmMotorLevy, 0);
-}
-
-void pravyMotorVpred(int rychlost) {
-  digitalWrite(inMotorPravy1, LOW);
-  digitalWrite(inMotorPravy2, HIGH);
-  analogWrite(pwmMotorPravy, rychlost);
-}
-
-void pravyMotorVzad(int rychlost) {
-  digitalWrite(inMotorPravy1, HIGH);
-  digitalWrite(inMotorPravy2, LOW);
-  analogWrite(pwmMotorPravy, rychlost);
-}
-
-void pravyMotorStop() {
-  analogWrite(pwmMotorPravy, 0);
-}
-
-
-void pohyb(int rychlostL, int rychlostR){ // doleva - levy opacny
-                                          // doprava - pravy opacny
-  if(rychlostL < 0){
-    levyMotorVzad(abs(rychlostL));
-    }
-  else{
-    levyMotorVpred(rychlostL);
-    }
-
-  if(rychlostR < 0){
-    pravyMotorVzad(abs(rychlostR));
-    }
-  else{
-    pravyMotorVpred(rychlostR);
-    }
-}
 
 /*
 int offset = 0;
@@ -295,31 +208,6 @@ void setup() {
 
 }
 
-void otacej_dokud_nenajdes_caru(byte position){ 
-  // 1001 jsme na care
-  // 1111 jsme mimo caru
-  // 0000 krizovatka
-  // byte cara = 0b1001;
-  // obsolete => otacej dle offsetu 
-  boolean leva = (position & 0b01000); 
-  boolean stred1 = (position & 0b00100);
-  boolean stred2 = (position & 0b00010);
-  boolean prava = (position & 0b00001);
-  if ( leva & !stred1 & !stred2 & prava ){
-    pohyb(0,0);
-  } else {
-    pohyb(-120, 120);
-  }
-}
-
-void otacej_dle_offsetu(int offset){
-  if ( (offset < 10) & (-10 < offset)){
-    pohyb(0,0);
-  } else {
-    pohyb(-120, 120);
-  }
-
-}
 
 void svit(byte position){
   if(0b00001000 & position){
